Migrate Typesense to IndexerInterface and SearcherInterface
