+++ packages/seal-typesense-adapter/Tests/TypesenseSearcherTest.php
@@ -3,9 +3,9 @@
 namespace Schranz\Search\SEAL\Adapter\Typesense\Tests;
 
 use Schranz\Search\SEAL\Adapter\Typesense\TypesenseAdapter;
-use Schranz\Search\SEAL\Testing\AbstractAdapterTestCase;
+use Schranz\Search\SEAL\Testing\AbstractSearcherTestCase;
 
-class TypesenseAdapterTest extends AbstractAdapterTestCase
+class TypesenseSearcherTest extends AbstractSearcherTestCase
 {
     public static function setUpBeforeClass(): void
     {
@@ -13,5 +13,10 @@ class TypesenseAdapterTest extends AbstractAdapterTestCase
         self::$adapter = new TypesenseAdapter($client);
 
         parent::setUpBeforeClass();
+    }
+
+    public function testFindMultipleIndexes(): void
+    {
+        $this->markTestSkipped('Not supported by Typesense: https://github.com/schranz-search/schranz-search/issues/98');
     }
 }
