+++ packages/seal-typesense-adapter/TypesenseSearcher.php
@@ -2,19 +2,16 @@
 
 namespace Schranz\Search\SEAL\Adapter\Typesense;
 
-use Schranz\Search\SEAL\Task\SyncTask;
+use Schranz\Search\SEAL\Adapter\SearcherInterface;
 use Typesense\Client;
-use Schranz\Search\SEAL\Adapter\ConnectionInterface;
 use Schranz\Search\SEAL\Marshaller\Marshaller;
 use Schranz\Search\SEAL\Schema\Index;
 use Schranz\Search\SEAL\Search\Condition;
 use Schranz\Search\SEAL\Search\Result;
 use Schranz\Search\SEAL\Search\Search;
-use Schranz\Search\SEAL\Task\TaskInterface;
 use Typesense\Exceptions\ObjectNotFound;
-use Typesense\Exceptions\RequestMalformed;
 
-final class TypesenseConnection implements ConnectionInterface
+final class TypesenseSearcher implements SearcherInterface
 {
     private Marshaller $marshaller;
 
@@ -23,37 +20,6 @@ final class TypesenseConnection implements ConnectionInterface
     ) {
         $this->marshaller = new Marshaller(dateAsInteger: true);
     }
-
-    public function save(Index $index, array $document, array $options = []): ?TaskInterface
-    {
-        $identifierField = $index->getIdentifierField();
-
-        /** @var string|null $identifier */
-        $identifier = ((string) $document[$identifierField->name]) ?? null;
-
-        $marshalledDocument = $this->marshaller->marshall($index->fields, $document);
-        $marshalledDocument['id'] = $identifier;
-
-        $this->client->collections[$index->name]->documents->upsert($marshalledDocument);
-
-        if (true !== ($options['return_slow_promise_result'] ?? false)) {
-            return null;
-        }
-
-        return new SyncTask(null);
-    }
-
-    public function delete(Index $index, string $identifier, array $options = []): ?TaskInterface
-    {
-        $this->client->collections[$index->name]->documents[$identifier]->delete();
-
-        if (true !== ($options['return_slow_promise_result'] ?? false)) {
-            return null;
-        }
-
-        return new SyncTask(null);
-    }
-
     public function search(Search $search): Result
     {
         // optimized single document query
